Функции сохранения настроек вынесены в отдельный файл
