+++ myClient/settings.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "mainwindow.h"
+
+
